fix(vrack-services): fix playwright tests
ref: MANAGER-0000
