+++ packages/manager/apps/vrack-services/cucumber.js
@@ -3,7 +3,6 @@ module.exports = {
     paths: ['e2e/features/**/*.feature'],
     require: [
       '../../../../playwright-helpers/bdd-setup.ts',
-      './mock/bdd-mock-setup.ts',
       'e2e/**/*.step.ts',
     ],
     requireModule: ['ts-node/register'],
