+++ packages/manager/modules/order/package.json
@@ -1,17 +1,16 @@
 {
-  "name": "@ovh-ux/manager-core-order",
+  "name": "@ovh-ux/manager-module-order",
   "version": "0.1.0",
   "private": true,
-  "description": "OVHcloud manager core order.",
+  "description": "OVHcloud manager module order.",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ovh/manager.git",
-    "directory": "packages/manager/core/order"
+    "directory": "packages/manager/modules/order"
   },
   "license": "BSD-3-Clause",
   "author": "OVH SAS",
   "sideEffects": false,
-  "type": "module",
   "main": "./dist/index.js",
   "module": "./dist/index.js",
   "types": "./dist/types/index.d.ts",
@@ -24,12 +23,12 @@
     "start:watch": "tsc -w"
   },
   "dependencies": {
-    "@types/jsurl": "^0.1.2",
     "@ovh-ux/manager-core-api": "^0.3.0",
     "@ovh-ux/manager-react-shell-client": "^0.1.0",
     "@tanstack/react-query": "^5.18.1",
-    "react": "^18.2.0",
-    "jsurl": "^0.1.5"
+    "@types/jsurl": "^0.1.2",
+    "jsurl": "^0.1.5",
+    "react": "^18.2.0"
   },
   "devDependencies": {
     "typescript": "^4.3.2"
